requirements.txt
crewai==0.126.0
fastapi==0.115.12
matplotlib==3.10.3
numpy==2.3.0
pandas==2.3.0
python-dotenv==1.1.0
requests==2.32.4
streamlit==1.45.1
uvicorn==0.34.3
